// frontend/features/api/model/shiftIndex200.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Petrol Integrity System API Documentation
 * OpenAPI spec version: 0.0.1
 */
import type { ShiftIndex200Links } from './shiftIndex200Links';
import type { ShiftIndex200Meta } from './shiftIndex200Meta';
import type { ShiftResource } from './shiftResource';

export type ShiftIndex200 = {
  data: ShiftResource[];
  links: ShiftIndex200Links;
  meta: ShiftIndex200Meta;
};