// frontend/features/api/model/shiftResource.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Petrol Integrity System API Documentation
 * OpenAPI spec version: 0.0.1
 */
import type { CreditSaleResource } from './creditSaleResource';
import type { DipReading } from './dipReading';
import type { MeterReading } from './meterReading';
import type { Payment } from './payment';
import type { ShiftResourceFinancials } from './shiftResourceFinancials';
import type { ShiftResourceWetStock } from './shiftResourceWetStock';

export interface ShiftResource {
  id: string;
  station_name: string;
  started_at: string;
  status: string;
  variance_alert: string;
  financials: ShiftResourceFinancials;
  wet_stock: ShiftResourceWetStock;
  readings?: MeterReading[];
  dips?: DipReading[];
  payments?: Payment[];
  credit_sales?: CreditSaleResource[];
}